refactor(ui): split dashboard.rs monolith into pages/dashboard/ module
Pure refactor — 1544-line dashboard.rs becomes a 10-file module
with one concern per file (card_panel, pass_banner, charge_section,
topup_section, transactions_list, block_button, edit_info_form,
sell_pass_modal, helpers). No behavioural changes, all data-testids
and classnames preserved. Later tasks apply new design primitives
and convert modals to sheets.
